Add files via upload
Main changes: 
Added an offline (non-bot connected) main, which means extra maintenance but doesn't necessarily need this maintenance.
Changed how functions are... functioning. Now mainly using [create, do, view] so that things are uniform. also main bot now has "admin" commands.
Finished the move function, it is part of the action file, but may be moved to the map file so that there are less passing arguments (the action debate)
Currently avoiding structural decisions and going to begin work on enemy entity so that I can begin work to test a mutator