Add files via upload
Changed the structure of the code to include folders, as the amount of files were being quite heavy. Fixed a bug related to the "view character short" function since the actions changed. Added a parameter to map creation to streamline further map testing. Added "entityList" to map.py in order to make finding entities on the map faster (previously was iterating the entire map), additionally added a position attribute to all entities.